Pulse: fix type for format string
Pointed by clang.

Reviewed-on: https://review.haiku-os.org/c/haiku/+/2405

+++ src/apps/pulse/NormalPulseView.cpp
@@ -80,7 +80,7 @@ NormalPulseView::NormalPulseView(BRect rect)
 			CPUBUTTON_MLEFT + CPUBUTTON_WIDTH + 7,
 			CPUBUTTON_MTOP + ITEM_OFFSET * x + CPUBUTTON_HEIGHT + 7);
 		char temp[4];
-		snprintf(temp, sizeof(temp), "%hhd", x + 1);
+		snprintf(temp, sizeof(temp), "%hhd", int8(x + 1));
 		fCpuButtons[x] = new CPUButton(r, B_TRANSLATE("Pulse"), temp, NULL);
 		AddChild(fCpuButtons[x]);
 	}
